requirements.txt
PyMuPDF==1.24.9
python-docx==1.1.2
edge-tts==6.1.12
pydub==0.25.1
pygame==2.6.1
tk==0.1.0
tkinter-tooltip==3.0.1
requests==2.32.3